Practice/src/test/java/org/example/AppTest.java
package org.example;

import static org.junit.Assert.assertEquals;
import org.junit.Test;
import junit.framework.TestCase;
import junit.framework.TestSuite;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;

/**
 * Unit test for simple App.
 */
@RunWith(SpringJUnit4ClassRunner.class)
public class AppTest extends TestCase {
    /**
     * Create the test case
     *
     * @param testName name of the test case
     */
    public AppTest(String testName) {
        super(testName);
    }

    /**
     * @return the suite of tests being tested
     */
    public static Test suite() {
        return new TestSuite(AppTest.class);
    }

    /**
     * Rigorous Test :-)
     */
    public void testApp() {
        assertTrue(true);
    }

   // @autowired
    App app;

    @Test
    public void getC() {
        int a = 10;
        int b = 20;

        int actualResult = app.getC(a, b);
        int expectedResult = 30;

        assertEquals(expectedResult, actualResult);
    }
}
